requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.5.2
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
autopep8==2.0.0
blis==0.7.9
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.9.24
charset-normalizer==2.1.1
ckiptagger==0.2.1
click==8.1.3
confection==0.0.3
cymem==2.0.7
Django==3.2.16
flatbuffers==22.12.6
frozenlist==1.3.3
future==0.18.2
gast==0.4.0
gensim==4.2.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
idna==3.4
importlib-metadata==5.1.0
Jinja2==3.1.2
keras==2.11.0
langcodes==3.3.0
libclang==14.0.6
line-bot-sdk==2.3.0
Markdown==3.4.1
MarkupSafe==2.1.1
multidict==6.0.3
murmurhash==1.0.9
numpy==1.21.6
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==22.0
pandas==1.3.5
pathy==0.10.1
preshed==3.0.8
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pydantic==1.10.2
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.7.3
six==1.16.0
smart-open==6.3.0
spacy==3.4.4
spacy-legacy==3.0.10
spacy-loggers==1.0.4
spacy-pkuseg==0.0.32
sqlparse==0.4.3
srsly==2.4.5
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
thinc==8.1.5
tomli==2.0.1
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.1.1
urllib3==1.26.13
wasabi==0.10.1
Werkzeug==2.2.2
wrapt==1.14.1
yarl==1.8.2
zh-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/zh_core_web_sm-3.4.0/zh_core_web_sm-3.4.0-py3-none-any.whl
zipp==3.11.0
